fix/ai tools

+++ apps/user/desktop/tool/browser/vivaldi/default.nix
@@ -45,16 +45,18 @@ in
     };
     activation.applyVivaldiConfig = hm.dag.entryAfter [ "writeBoundary" ] (
       optionalString pkgs.stdenv.hostPlatform.isDarwin ''
-        find "$HOME/Library/Application Support/Vivaldi" -maxdepth 1 -type d -name "Default" -or -name "Profile *" | while read -r profile; do
-          TMP="''${profile}/Preferences.bak"
-          mv "''${profile}/Preferences" "''${TMP}"
-          ${pkgs.jq}/bin/jq -r -s '.[0] * .[1]' "''${TMP}" ${getChezmoiFilePath "dot_config/vivaldi/CommonPreferences"} > "''${profile}/Preferences.new"
-          if [ -s "''${profile}/Preferences.new" ]; then
-            mv "''${profile}/Preferences.new" "''${profile}/Preferences"
-          else
-            rm "''${profile}/Preferences.new"
-          fi
-        done
+        if [-d "$HOME/Library/Application Support/Vivaldi" ]; then
+          find "$HOME/Library/Application Support/Vivaldi" -maxdepth 1 -type d -name "Default" -or -name "Profile *" | while read -r profile; do
+            TMP="''${profile}/Preferences.bak"
+            mv "''${profile}/Preferences" "''${TMP}"
+            ${pkgs.jq}/bin/jq -r -s '.[0] * .[1]' "''${TMP}" ${getChezmoiFilePath "dot_config/vivaldi/CommonPreferences"} > "''${profile}/Preferences.new"
+            if [ -s "''${profile}/Preferences.new" ]; then
+              mv "''${profile}/Preferences.new" "''${profile}/Preferences"
+            else
+              rm "''${profile}/Preferences.new"
+            fi
+          done
+        fi
       ''
     );
   };
